refactor(hub): repoint importers to kernel/-relocated seams (stage 3e)
Rewrite every internal importer of the four seams to the kernel/ location
across src/ and __tests__/ (src root ./seam -> ./kernel/seam; kernel siblings
../seam -> ./seam; with-* ../../seam -> ../../kernel/seam; tests ../src/seam ->
../src/kernel/seam). Golden adapter-surface test reads the new source path; its
frozen symbol list is unchanged. Behavior-preserving; 2974 tests green.
