+++ packages/hub/__tests__/attachments.test.ts
@@ -273,7 +273,7 @@ describe('wrapBundleStore', () => {
   it('wraps a bundle into a full NoydbStore', async () => {
     const storage = new Map<string, { bytes: Uint8Array; version: string }>()
     let versionCounter = 0
-    const { wrapBundleStore } = await import('../src/store/bundle-store.js')
+    const { wrapBundleStore } = await import('../src/kernel/store/bundle-store.js')
 
     const bundleStore = wrapBundleStore({
       kind: 'bundle',
@@ -324,7 +324,7 @@ describe('wrapBundleStore', () => {
   it('conflict check: expectedVersion throws ConflictError on mismatch', async () => {
     const storage = new Map<string, { bytes: Uint8Array; version: string }>()
     let versionCounter = 0
-    const { wrapBundleStore } = await import('../src/store/bundle-store.js')
+    const { wrapBundleStore } = await import('../src/kernel/store/bundle-store.js')
 
     const bundleStore = wrapBundleStore({
       kind: 'bundle',
