+++ packages/hub/__tests__/query-join-multi.test.ts
@@ -15,7 +15,7 @@ import type {
   VaultSnapshot,
 } from '../src/kernel/types.js'
 import { ConflictError, JoinTooLargeError } from '../src/kernel/errors.js'
-import { resetJoinWarnings } from '../src/query/index.js'
+import { resetJoinWarnings } from '../src/kernel/query/index.js'
 import { ref } from '../src/kernel/refs.js'
 
 /** Same memory adapter shape used in query-integration / query-join tests. */
