+++ packages/hub/src/kernel/collection-config.ts
@@ -46,7 +46,7 @@ import type { ComputedFields } from '../with-formula/computed/index.js'
 import type { FieldMeta } from '../with-shape/introspection/field-meta.js'
 import type { CollectionMeta } from '../with-shape/introspection/meta.js'
 import type { RefDescriptor } from './refs.js'
-import type { JoinableSource } from '../query/index.js'
+import type { JoinableSource } from './query/index.js'
 import { VectorSet, type EmbeddingDescriptor } from '../with-lookup/embeddings/index.js'
 import { Lru } from './cache/index.js'
 import type { ReadOnlyVaultFacade } from '../with-audit/guards/types.js'
