+++ packages/hub/src/kernel/collection.ts
@@ -67,8 +67,8 @@ import { derivePersistedSchema } from '../with-shape/persisted-schemas/derive.js
 import type { LedgerStore } from '../with-commit/history/ledger/index.js'
 import type { DiffEntry } from '../with-commit/history/diff.js'
 import type { HistoryStrategy } from '../with-commit/history/strategy.js'
-import { Query, ScanBuilder } from '../query/index.js'
-import type { QuerySource, JoinContext, JoinableSource } from '../query/index.js'
+import { Query, ScanBuilder } from './query/index.js'
+import type { QuerySource, JoinContext, JoinableSource } from './query/index.js'
 import type { CollectionIndexes } from '../with-lookup/indexing/eager-indexes.js'
 import { decodeIdxId } from '../with-lookup/indexing/persisted-indexes.js'
 import type { PersistedCollectionIndex } from '../with-lookup/indexing/persisted-indexes.js'
