+++ packages/hub/src/kernel/noydb.ts
@@ -95,8 +95,8 @@ import type { DryRunResult } from '../with-commit/tx/dry-run.js'
 import { NO_TX, type TxStrategy } from '../with-commit/tx/strategy.js'
 import { NO_FORGET, type ForgetStrategy } from '../with-audit/forget/strategy.js'
 import { readDottedPath, coerceSubjectId } from '../with-audit/forget/subject-index.js'
-import { INDEXED_STORE_POLICY } from '../store/sync-policy.js'
-import { memoryStore } from '../store/memory-store.js'
+import { INDEXED_STORE_POLICY } from './store/sync-policy.js'
+import { memoryStore } from './store/memory-store.js'
 import type { PolicyEnforcer } from '../with-party/session/session-policy.js'
 import { NO_SESSION, type SessionStrategy } from '../with-party/session/strategy.js'
 import {
