+++ packages/hub/src/kernel/types.ts
@@ -25,7 +25,7 @@
 
 import type { StandardSchemaV1 } from './schema.js'
 import type { DeferredNumberingConfig } from '../with-commit/numbering/descriptor.js'
-import type { SyncPolicy } from '../store/sync-policy.js'
+import type { SyncPolicy } from './store/sync-policy.js'
 import type { BlobStrategy } from '../with-shape/blobs/strategy.js'
 import type { ArchiveStrategy } from '../with-fork/archive/index.js'
 import type { IndexStrategy } from '../with-lookup/indexing/strategy.js'
