+++ packages/hub/src/kernel/vault.ts
@@ -23,7 +23,7 @@ import { NOYDB_FORMAT_VERSION } from './types.js'
 import { Collection } from './collection.js'
 import type { CacheOptions } from './collection.js'
 import type { IndexDef } from '../with-lookup/indexing/eager-indexes.js'
-import type { JoinableSource } from '../query/index.js'
+import type { JoinableSource } from './query/index.js'
 import type { OnDirtyCallback } from './collection.js'
 import type { UnlockedKeyring, BundleRecipient } from '../with-party/team/keyring.js'
 import { exportAccessibleData } from '../with-audit/portability/export-accessible.js'
