+++ packages/hub/src/with-formula/materialized-views/registry.ts
@@ -1,7 +1,7 @@
 import { MaterializedViewCycleError, MaterializedViewSourceUnknownError } from '../../kernel/errors.js'
 import type { DerivationRegistry } from '../derivations/registry.js'
-import type { Clause, FieldClause } from '../../query/predicate.js'
-import type { DeclaredPredicate } from '../../query/builder.js'
+import type { Clause, FieldClause } from '../../kernel/query/predicate.js'
+import type { DeclaredPredicate } from '../../kernel/query/builder.js'
 import { analyzeDependencies, summarizeQueryPlan, summarizeUnionPlan } from './dependency-analyzer.js'
 import { computeQueryHash } from './query-hash.js'
 import type { MaterializedViewStrategy, MVQueryContext } from './types.js'
