+++ packages/hub/src/with-lookup/indexing/lazy-builder.ts
@@ -23,8 +23,8 @@
  * indexes exist to prevent (see `docs/architecture.md` §indexes).
  */
 
-import type { Clause, FieldClause, Operator } from '../../query/predicate.js'
-import { evaluateClause, readPath } from '../../query/predicate.js'
+import type { Clause, FieldClause, Operator } from '../../kernel/query/predicate.js'
+import { evaluateClause, readPath } from '../../kernel/query/predicate.js'
 import type { PersistedCollectionIndex } from './persisted-indexes.js'
 import { IndexRequiredError } from '../../kernel/errors.js'
 import type { QueryField } from '../../kernel/types.js'
