+++ packages/hub/src/with-lookup/indexing/unique-constraints.ts
@@ -15,7 +15,7 @@
  * throw at registration instead (see Collection constructor).
  */
 
-import { readPath } from '../../query/predicate.js'
+import { readPath } from '../../kernel/query/predicate.js'
 import { canonicalGroupKey } from '../aggregate/canonical-key.js'
 import { UniqueConstraintError, UnsupportedIndexOptionError } from '../../kernel/errors.js'
 import type { IndexDef } from './eager-indexes.js'
